Move edb examples
